refactor(front): Pass A billing/credits/spend gotchas — move DTOs + pure local types
Pass A: for gotcha files where a DTO references a same-file domain type, move the
DTO plus the pure local type to types/api; defer files whose local type
transitively couples to @app/lib.

Migrated (7): billing/info, billing/invoices, credits/metronome_contract,
credits/usage_configuration, subscription, users/spend_limit,
workspace/default_user_spend_limit. Pure domain types (BillingInfo,
BillingInvoice, MetronomeContractSummary, CreditUsageConfigurationBody,
CheckoutStatus, UserSpendLimit, DefaultUserSpendLimit, ...) moved alongside
their DTOs; lib back-imports the ones returned by kept functions.

Deferred to Pass B (transitive @app/lib couplings): credits/members_usage
(BillingFrequency/EffectiveSpendLimitSource/MetronomeAlertRef),
credits/purchase (CreditPurchaseLimits), credits/seat_plan (SeatAwuCreditsPeriod).

21 consumers updated (split-aware).
