+++ front-api/routes/w/[wId]/subscriptions/checkout-status.ts
@@ -1,4 +1,4 @@
-import type { GetCheckoutStatusResponseBody } from "@app/lib/api/subscription";
+import type { GetCheckoutStatusResponseBody } from "@app/types/api/subscription";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import type { HandlerResult } from "@front-api/middlewares/utils";
