+++ front-api/routes/w/[wId]/subscriptions/index.ts
@@ -1,11 +1,4 @@
-import type {
-  GetSubscriptionsResponseBody,
-  PostSubscriptionResponseBody,
-} from "@app/lib/api/subscription";
-import {
-  isMetronomeBillingEnabled,
-  PatchSubscriptionRequestBody,
-} from "@app/lib/api/subscription";
+import { isMetronomeBillingEnabled } from "@app/lib/api/subscription";
 import {
   createCheckoutUrl,
   PostSubscriptionRequestBody,
@@ -17,6 +10,11 @@ import {
 } from "@app/lib/plans/stripe";
 import { SubscriptionResource } from "@app/lib/resources/subscription_resource";
 import logger from "@app/logger/logger";
+import type {
+  GetSubscriptionsResponseBody,
+  PostSubscriptionResponseBody,
+} from "@app/types/api/subscription";
+import { PatchSubscriptionRequestBody } from "@app/types/api/subscription";
 import { assertNever } from "@app/types/shared/utils/assert_never";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import {
