+++ front/types/api/users/spend_limit.ts
@@ -1,0 +1,13 @@
+export type UserSpendLimit =
+  | { kind: "unlimited" }
+  | { kind: "limited"; awuCredits: number };
+
+export type GetUserSpendLimitResponse = UserSpendLimit;
+
+export type GetUserSpendLimitResponseBody = GetUserSpendLimitResponse;
+
+export type PutUserSpendLimitResponseBody = SetUserSpendLimitResponse;
+
+export type SetUserSpendLimitResponse = {
+  limit: UserSpendLimit;
+};
