+++ front/types/api/workspace/default_user_spend_limit.ts
@@ -1,0 +1,9 @@
+export type DefaultUserSpendLimit = {
+  awuCredits: number;
+};
+
+export type GetDefaultUserSpendLimitResponseBody = {
+  awuCredits: number | null;
+};
+
+export type PutDefaultUserSpendLimitResponseBody = DefaultUserSpendLimit;
